PROTOCOL CHANGE: send sever rank as a float
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18514 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDRenderer.cxx
@@ -1138,7 +1138,7 @@ void			HUDRenderer::renderStatus(void)
   fm.drawString(x, y, 0, majorFontFace->getFMFace(), majorFontSize, buffer);
 }
 
-int HUDRenderer::tankScoreCompare(const void* _a, const void* _b)
+float HUDRenderer::tankScoreCompare(const void* _a, const void* _b)
 {
   World *world = World::getWorld();
   if (!world) {
