+++ src/bzflag/HUDRenderer.h
@@ -164,7 +164,7 @@ private:
 
   void		resize(bool firstTime);
   static void	resizeCallback(void*);
-  static int	tankScoreCompare(const void* _a, const void* _b);
+  static float	tankScoreCompare(const void* _a, const void* _b);
   static int	teamScoreCompare(const void* _a, const void* _b);
 
   GLDisplayList	friendlyMarkerList;
